made it a bit more ADOM-like
Walls are #'s and the player is an @ symbol. Changed the warp to a W, makes more sense to me, idk. Trying to make it so we can have moving enemies too.

If we're thinking of turning this into a roguelike though, we may as well use a library for this stuff lol. We'll see what happens