Made a Representation class, moved all chars and color to that class
I'm not sure if it was a good idea yet or not, but this way I can reuse
representations of stuff. I started doing it because I wanted to define a
universal 'dead' representation, but thinknig about it, other than that,
there's not much shared representation stuff, because now there's going to
be literally double the classes and files I'll have to be writing.

We'll see how it goes. Trying harder to just write out the code and worry
about how it works later because it's too tempting for me to constanly be
focusing on the future. That'll come later when I'm actually experienced
enough to figure out what needs to come later.