Sloppily forced a master member into inventory
That way a tile can still have an inventory while most actors can too, and
have the convenience of having access to the Actor class. I feel like this
is part of a design error though, where objects are taking too much inputs
only to maybe be using parts of them. I'm not sure though, but I'm sure
I'll learn more about it as I go.