no longer using std namespace in combat, explode on levelup
need to sort out some sort of event driven ui for drawing. can't keep
setting flags. It'd be a bit easier to create some sort of event queue
where I define an event and then a handler for it, along with frames of
animations and stuff. I dunno how exactly but it's going to have to happen
if I want to make my game look good.