Refactored tiletypes to use less memory
A tile has the id and it uses a pseudo-static class of BaseTileType or its
derivatives and then makes a copy if it needs to change it.

disabled mouse highlighting and level up animations  because I didn't want
to worry about temp colors, I'll need to rework that later

also disabled grants warp tiles and build mode. they had a good run,
thanks for all the help dude!